h3: initial version of extensible priority scheme
This builds on top of the QUIC-level priority support.

This adds a Priority struct used to hold Extensibile priority
along with methods to convert them to/from wire.

Internally every stream is now marked with an Extensible Priority.
Importantly, this affects how HTTP response body bytes will be
scheduled when flushing streams. The HTTP/3 API is updated with
the new method `response_priority()` that servers can use to set
the priority of the HTTP response. The chosen priority value is
independent of any received client signal. At this time the
priority cannot be changed after HEADERS have been written via
`send_response()`. Future work is needed to support changing the
priority of in-flight responses, but that should not affect the
API.

No new API method is provided for clients to set request priority.
However, the `Priority` HTTP header can be provided in the
`send_request()` method.

The quiche-client and quiche-server applications have been updated
to test this out. quiche-client can be run with e.g.
`-H "Priority:u=4,i"` to send a priority signal. quiche-server
process the Priority header (or assumes the default if none is
provided). It also supports a special query string syntax that
takes precedence over the header. For example, a client can send
`https://example.com?u=4&i=1` to indicate the priority as `u=1,i`.