retransmit MAX_DATA when lost
Same idea of #525 but we need a little change because
after we normally send MAX_DATA the next call of
should_update_max_data() returns false so no way to
trigger retransmit.

The change is to have a similar way of updating
MAX_STREAM_DATA. The connection is marked as full and
trigger to send MAX_DATA when self.almost_full is true.